Add Github Action workflow
The workflow runs two separate phases:
1. Generate the build matrix
2. Build the container images according to the build matrix

The build matrix generation is done by getting all the kernel versions
for the Red Hat CoreOS (RHCOS) images. Each OpenShift release comes with a new
build of RHCOS. However, the kernel version is the same for many RHCOS
builds, so script deduplicates the kernel versions to reduce the build
matrix size. The build matrix also list the available architectures for
a given kernel version, allowing for multi-arch builds.

Then, the build phase uses the matrix to trigger parallel jobs. Each job
is responsible for a kernel version and support multi-arch. To avoid
rebuilding existing images, the job checks if the image manifest exists
in the target registry and skips the build if that is the case.
Otherwise, it uses the `docker/build-push-action` action to build and
push the container image.

The workflow is configured to be triggered every day at midnight or
manually.

Signed-off-by: Fabien Dupont <fabiendupont@pm.me>