Add a build condition on UBI digest (#4)
The `driver-toolkit` image is based on `ubi8/ubi` and we want to ensure
that is rebuilt when the parent image is updated. The minor release tag
is a floating tag and will point to a new image after an async update.

We store the base image digest in the
`org.opencontainers.image.base.digest` label of the final image, in
order to compare it with the `ubi8/ubi` digest for the floating tag.

If the digests don't match, it means that an updated base image is
available and we trigger and build/push of the `driver-toolkit` image.

Signed-off-by: Fabien Dupont <fabiendupont@pm.me>